Revert the second change
After a few test, I found this way may cause sequence asymmetry between angle and torsion. The two of them may not be the same length and there are many zeros in the new version torsion. To deal with this problem may add more computational time. So I revert the second change.